Reorganize backend code. Delete unused files. Switch to absolute imports
